Remove error message when ./empty_trash.sh is trying to clean empty log folder

+++ main/scripts/empty_trash.sh
@@ -1,2 +1,6 @@
 #!/bin/bash
-rm ../trash/nohup_output_*.txt
+
+if [ -e ../trash/nohup_output_*.txt ]
+then
+    rm ../trash/nohup_output_*.txt
+fi
